+++ src/sql/src/func.rs
@@ -2184,7 +2184,7 @@ lazy_static! {
             // Aggregates.
             "array_agg" => Aggregate {
                 params!(NonVecAny) => Operation::unary_ordered(|ecx, e, order_by| {
-                    if let ScalarType::Char {.. }  = ecx.scalar_type(&e) {
+                    if let ScalarType::Char {.. } | ScalarType::List | ScalarType::Map  = ecx.scalar_type(&e) {
                         bail_unsupported!("array_agg on char");
                     };
                     // ArrayConcat excepts all inputs to be arrays, so wrap all input datums into